feat: make fold 2 and fold 3 ssr
